# smart-portfolio-main/backend/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Smart Portfolio API
  description: |
    REST API for a personal developer portfolio with AI-powered resume chat (RAG),
    contact form with Discord notifications, project showcase management, and
    Razorpay sponsorship payments.

    ## Authentication

    Admin-protected endpoints require one of:
    - `X-Admin-Key` header with the configured `ADMIN_API_KEY`
    - `Authorization: Bearer <ADMIN_API_KEY>` header

    When `ADMIN_API_KEY` is not set, admin endpoints are accessible without authentication
    (suitable for local development only).

    ## Response Envelope

    Every endpoint returns a consistent JSON envelope:

    ```json
    {
      "success": true,
      "data": { ... }
    }
    ```

    Error responses:

    ```json
    {
      "success": false,
      "error": {
        "code": 400,
        "message": "title is required; description is required"
      }
    }
    ```

    ## SSE Streaming

    `POST /api/chat/stream` returns a `text/event-stream` response.
    Each LLM token is sent as `data: <token>\n\n`.
    A final `event: done\ndata: \n\n` frame signals stream completion.
  version: "1.0.0"
  contact:
    name: ZRishu
    url: https://github.com/ZRishu/smart-portfolio
  license:
    name: MIT
...
components:
  securitySchemes:
    AdminApiKey:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: Admin API key passed via the X-Admin-Key header
    BearerAuth:
      type: http
      scheme: bearer
      description: Admin API key passed as a Bearer token
  schemas:
    APIResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          description: Response payload (present on success)
        error:
          $ref: "#/components/schemas/APIError"
    APIError:
      type: object
      properties:
        code:
          type: integer
          example: 400
        message:
          type: string
          example: "title is required"
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "d290f1ee-6c54-4b01-90e6-d701748f0851"
        title:
          type: string
          example: "Smart Portfolio"
        description:
          type: string
          example: "A full-stack portfolio with AI chat"
        tech_stack:
          type: string
          nullable: true
          example: "Go, PostgreSQL, pgvector"
        github_url:
          type: string
          nullable: true
          example: "https://github.com/ZRishu/smart-portfolio"
        live_url:
          type: string
          nullable: true
          example: "https://portfolio.example.com"
        created_at:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - created_at
    ProjectRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          example: "Smart Portfolio"
        description:
          type: string
          minLength: 1
          example: "A full-stack portfolio with AI chat"
        tech_stack:
          type: string
          nullable: true
          example: "Go, PostgreSQL, pgvector"
        github_url:
          type: string
          nullable: true
          example: "https://github.com/ZRishu/smart-portfolio"
        live_url:
          type: string
          nullable: true
          example: "https://portfolio.example.com"
      required:
        - title
        - description
    ContactMessage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sender_name:
          type: string
          example: "Jane Doe"
        sender_email:
          type: string
          format: email
          example: "jane@example.com"
        message_body:
          type: string
          example: "Hello! I loved your portfolio."
        is_read:
          type: boolean
        submitted_at:
          type: string
          format: date-time
    ContactMessageRequest:
      type: object
      properties:
        sender_name:
          type: string
          minLength: 1
          example: "Jane Doe"
        sender_email:
          type: string
          format: email
          example: "jane@example.com"
        message_body:
          type: string
          minLength: 1
          example: "Hello! I loved your portfolio."
      required:
        - sender_name
        - sender_email
        - message_body
    ContactMessageResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sender_name:
          type: string
        submitted_at:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Validation error or malformed request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIResponse"
          example:
            success: false
            error:
              code: 400
              message: "title is required; description is required"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIResponse"
          example:
            success: false
            error:
              code: 404
              message: "project not found"
    Unauthorized:
      description: Missing authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIResponse"
          example:
            success: false
            error:
              code: 401
              message: "authentication required — provide X-Admin-Key header or Authorization: Bearer <key>"
    Forbidden:
      description: Invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIResponse"
          example:
            success: false
            error:
              code: 403
              message: "invalid API key"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIResponse"
          example:
            success: false
            error:
              code: 500
              message: "an internal error occurred — please try again later"
paths:
...
  /api/projects:
    get:
      tags: [Projects]
      summary: List all projects
      description: Returns every portfolio project ordered by creation date descending. Results are served from an in-memory cache when available.
      operationId: listProjects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Project"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags: [Projects]
      summary: Create a project
      description: Creates a new portfolio project. Invalidates the project list cache.
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/projects/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Project UUID

    get:
      tags: [Projects]
      summary: Get a project by ID
      operationId: getProject
      responses:
        "200":
          description: Project found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      tags: [Projects]
      summary: Update a project
      description: Replaces all mutable fields of the project. Invalidates the project list cache.
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectRequest"
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags: [Projects]
      summary: Delete a project
      description: Permanently removes a project. Invalidates the project list cache.
      operationId: deleteProject
      responses:
        "204":
          description: Project deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/contact:
    post:
      tags: [Contact]
      summary: Submit a contact message
      description: |
        Persists a new contact form submission and fires an asynchronous Discord notification.
        This is a public endpoint — no authentication required.
      operationId: submitContact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactMessageRequest"
      responses:
        "201":
          description: Message submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/ContactMessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      tags: [Contact]
      summary: List all contact messages
      description: Returns every contact message ordered by submission date descending. Intended for admin dashboard use.
      operationId: listContactMessages
      security:
        - AdminApiKey: []
        - BearerAuth: []
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ContactMessage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/contact/unread:
    get:
      tags: [Contact]
      summary: List unread contact messages
      description: Returns only unread messages ordered by submission date descending.
      operationId: listUnreadContactMessages
      security:
        - AdminApiKey: []
        - BearerAuth: []
      responses:
        "200":
          description: List of unread messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ContactMessage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/contact/{id}/read:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Contact message UUID

    patch:
      tags: [Contact]
      summary: Mark a message as read
      operationId: markContactMessageRead
      security:
        - AdminApiKey: []
        - BearerAuth: []
      responses:
        "200":
          description: Message marked as read
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: "contact message marked as read"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/contact/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Contact message UUID

    delete:
      tags: [Contact]
      summary: Delete a contact message
      operationId: deleteContactMessage
      security:
        - AdminApiKey: []
        - BearerAuth: []
      responses:
        "204":
          description: Message deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...